Separated GraphQL errors from others
